refactor: Move date option generation into dateUtil
