+++ src/util/dateUtil.ts
@@ -7,3 +7,18 @@ export const isWeekend = (date: Date) => {
     return true;
   }
 };
+
+export const getNextSchooldays = (amount: number) => {
+  let genOptions = [];
+  let itDate = new Date();
+
+  while (genOptions.length < amount) {
+    if (!isWeekend(itDate)) {
+      genOptions.push(new Date(itDate));
+    }
+
+    itDate.setDate(itDate.getDate() + 1);
+  }
+
+  return genOptions;
+};
